Fix handling of defaults supplied at compile time.
Github issue #2: The refactoring of the config logic resulted in
the configure flags for various defaults being ignored. These flags
are now treated as default settings that can be overridden in the
suphp.conf file.

This change also corrects the default logfile path for the setuid
binary. This was listed as an option in "configure", but the path
was hardcoded in the code to a different location.